% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_email_usr.R
\name{html_msg}
\alias{html_msg}
\title{email_usr}
\usage{
html_msg(nome, email, msg, data, hora)
}
\value{
O valor ser retornado da função é um vetor carateres composto por um corpo do email e o texto que o usuário escreveu com seu nome e email.
}
\description{
A utils function para enviar mensagens personalizadas
}
